+++ src/main/java/miyucomics/hexical/features/prestidigitation/handlers/PrestidigitationHandlersEntity.kt
@@ -10,7 +10,10 @@ import net.minecraft.entity.TntEntity
 import net.minecraft.entity.decoration.ArmorStandEntity
 import net.minecraft.entity.mob.CreeperEntity
 import net.minecraft.entity.mob.EndermanEntity
-import net.minecraft.entity.passive.*
+import net.minecraft.entity.passive.LlamaEntity
+import net.minecraft.entity.passive.PandaEntity
+import net.minecraft.entity.passive.PufferfishEntity
+import net.minecraft.entity.passive.SquidEntity
 import net.minecraft.sound.SoundCategory
 import net.minecraft.sound.SoundEvents
 
@@ -46,7 +49,7 @@ object PrestidigitationHandlersEntity {
 		register(PrestidigitationHandlerEntity.simple(PandaEntity::class.java) { it.isSneezing = true })
 		register(PrestidigitationHandlerEntity.simple(SquidEntity::class.java, SquidEntity::squirt))
 		register(PrestidigitationHandlerEntity.simple(EndermanEntity::class.java) { env, enderman -> enderman.target = env.castingEntity })
-		register(PrestidigitationHandlerEntity.simple(FoxEntity::class.java, FoxEntity::spit))
+		//register(PrestidigitationHandlerEntity.simple(FoxEntity::class.java, FoxEntity::spit))
 
 		register(PrestidigitationHandlerEntity.simple(LlamaEntity::class.java) { env, llama ->
 			if (env.castingEntity != null)