+++ src/main/java/miyucomics/hexical/features/toast/OpSendTelepathy.kt
@@ -1,19 +1,18 @@
-package miyucomics.hexical.features.telepathy
+package miyucomics.hexical.features.toast
 
 import at.petrak.hexcasting.api.casting.castables.ConstMediaAction
 import at.petrak.hexcasting.api.casting.eval.CastingEnvironment
 import at.petrak.hexcasting.api.casting.iota.Iota
 import at.petrak.hexcasting.api.casting.mishaps.MishapBadCaster
-import net.minecraft.network.packet.s2c.play.TitleS2CPacket
 import net.minecraft.server.network.ServerPlayerEntity
 
-object OpShoutTelepathy : ConstMediaAction {
+object OpSendTelepathy : ConstMediaAction {
 	override val argc = 1
 	override fun execute(args: List<Iota>, env: CastingEnvironment): List<Iota> {
 		val caster = env.castingEntity
 		if (caster !is ServerPlayerEntity)
 			throw MishapBadCaster()
-		caster.networkHandler.sendPacket(TitleS2CPacket(args[0].display()))
+		caster.sendMessage(args[0].display(), true)
 		return emptyList()
 	}
 }